refactor(test runner): separated test runner to modules.
