+++ scripts/test_runner/lt_test_runner/adapter_mapping.csv
